+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 
 - Local Protocol :: Add lockAccount/unlockAccount method
 - Local Protocol :: Lock account if too many connection attempts have been sent on
+- Local Protocol :: Add more functional test
+
 
 ### v0.0.5 (2015/05/10)
 - Implement Local protocol